test(biblecore): migrate remote sync lifecycle tests
Why:
- Continue moving pure package-logic tests out of the app-host XCTest bundle so remote-sync contracts can run in the fast BibleCore package lane.
- Keep remote-sync lifecycle, adapter factory, background refresh, and reading-plan orchestration coverage while reducing simulator/app-host coupling.

What Changed:
- Moved `AndBibleTests+RemoteSyncLifecycle` into `BibleCoreTests` as `RemoteSyncLifecycleTests`.
- Added package-local lifecycle test support for synchronization reports, lifecycle synchronizer fakes, and iOS background refresh scheduler/task fakes.
- Removed the now-unused lifecycle-only fakes from app-host `AndBibleTestSupport` and removed stale Xcode project references to the migrated app-host file.
- Documented the migration-plan progress for the lifecycle slice.

Validation:
- `npx gitnexus analyze`
- GitNexus impact checks for `AndBibleTests+RemoteSyncLifecycle.swift` and `AndBibleTestSupport.swift`; both reported CRITICAL test/import wiring blast radius with 0 production execution flows.
- `gitnexus_detect_changes(scope: staged)` reported no production execution-flow impact.
- `plutil -lint AndBible.xcodeproj/project.pbxproj`
- `git diff --check` and `git diff --cached --check`
- `DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer swift package describe --type json`
- `python3 scripts/check_repo_standards.py docblocks --all-files`
- `python3 -m unittest discover -s scripts -p 'test_*.py'`
- `DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme BibleCoreTests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-remote-sync-lifecycle-final-dd CODE_SIGNING_ALLOWED=NO`
- `DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild build-for-testing -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-remote-sync-lifecycle-app-rerun-dd CODE_SIGNING_ALLOWED=NO`

Impact:
- No app runtime behavior changes.
- App-host unit bundle sheds another remote-sync logic file and stale lifecycle fakes.
- BibleCore package lane now owns 25 lifecycle tests, with iOS-only background refresh tests guarded for multi-platform package compilation.
