+++ docs/test-architecture-migration-plan.md
@@ -144,6 +144,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+RemoteSyncAdapters` and `AndBibleTests+RemoteSyncTransport` have moved to `BibleCoreTests` because they exercise WebDAV/Nextcloud transport request construction, multistatus parsing, and sync-folder marker behavior without UI or app bootstrap.
 - `AndBibleTests+RemoteSyncReadingPlans` has moved to `BibleCoreTests` because it exercises Android-compatible reading-plan status storage, snapshot restore, patch replay, and initial backup upload through BibleCore services without UI or app bootstrap.
 - `AndBibleTests+RemoteSyncState` has moved to `BibleCoreTests` because it exercises iCloud startup recovery, Android-compatible sync settings/state keys, WebDAV configuration, bootstrap coordination, patch discovery, and archive staging through BibleCore services without UI or app bootstrap.
+- `AndBibleTests+RemoteSyncLifecycle` has moved to `BibleCoreTests` because it exercises category synchronization lifecycle, WebDAV adapter factory selection, reading-plan patch orchestration, and background-refresh coordination through BibleCore services without UI or app bootstrap.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; `testContentViewDoesNotContainLegacyRootSidebarShell` becomes a repo-standards guardrail or stays app-hosted (decide explicitly - do not move to BibleUITests); the other ~78 funcs go to BibleUITests.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
